Make consistent --web flag to base groups
