+++ cli/nu.py
@@ -1,19 +1,19 @@
-import os.path
-import shutil
-
 import click
-from shellingham import detect_shell
 
 from cli.datasets import datasets
-from cli.docs import docs
+from cli.helpers.web_helper import launch_web_or_show_help
+from cli.install_completion import install_completion
 from cli.jobs import jobs
-from cli.modelci import modelci
 from cli.models import models
+from cli.reference import reference
 from cli.slices import slices
+from cli.unit_tests import unit_tests
 
 
-@click.group("cli")
-def nu():
+@click.group("cli", invoke_without_command=True)
+@click.option("--web", is_flag=True, help="Launch browser")
+@click.pass_context
+def nu(ctx, web):
     """Nucleus CLI (nu)
 
         \b
@@ -26,43 +26,17 @@ def nu():
 
         `nu` is a command line interface to interact with Scale Nucleus (https://dashboard.scale.com/nucleus)
     """
-    pass
+    launch_web_or_show_help(sub_url="", ctx=ctx, launch_browser=web)
 
 
-@nu.command("install-completion")
-def install_completion():
-    """Install shell completion script to your rc file"""
-    shell, _ = detect_shell()
-    if shell == "zsh":
-        rc_path = "~/.zshrc"
-        append_to_file = 'eval "$(_NU_COMPLETE=zsh_source nu)"'
-    elif shell == "bash":
-        rc_path = "~/.bashrc"
-        append_to_file = 'eval "$(_NU_COMPLETE=bash_source nu)"'
-    elif shell == "fish":
-        rc_path = "~/.config/fish/completions/foo-bar.fish"
-        append_to_file = "eval (env _NU_COMPLETE=fish_source nu)"
-    else:
-        raise RuntimeError(f"Unsupported shell {shell} for completions")
-
-    rc_path_expanded = os.path.expanduser(rc_path)
-    rc_bak = f"{rc_path_expanded}.bak"
-    shutil.copy(rc_path_expanded, rc_bak)
-    click.echo(f"Backed up {rc_path} to {rc_bak}")
-    with open(rc_path_expanded, mode="a") as rc_file:
-        rc_file.write("\n")
-        rc_file.write("# Shell completion for nu\n")
-        rc_file.write(append_to_file)
-    click.echo(f"Completion script added to {rc_path}")
-    click.echo(f"Don't forget to `source {rc_path}")
-
-
-nu.add_command(modelci)  # type: ignore
 nu.add_command(datasets)  # type: ignore
-nu.add_command(models)  # type: ignore
-nu.add_command(slices)  # type: ignore
-nu.add_command(docs)  # type: ignore
+nu.add_command(install_completion)  # type: ignore
 nu.add_command(jobs)  # type: ignore
+nu.add_command(models)  # type: ignore
+nu.add_command(reference)  # type: ignore
+nu.add_command(slices)  # type: ignore
+nu.add_command(unit_tests)  # type: ignore
 
 if __name__ == "__main__":
+    """To debug, run this script followed by request command tree e.g. `cli/nu.py datasets list`"""
     nu()
